[NETSPORTS-0] Added second y-axis to team performance chart. Changed opacity for bar charts.
