+++ dashboard.html
@@ -43,6 +43,6 @@
         </div>
     </div>
 
-    <script src="js/build_chart.js" type="text/javascript"></script>
+    <script type="module" src="js/data_preparation.js"></script>
 </body>
 </html>
