+++ js/build_chart.js
@@ -1,67 +1,4 @@
-function load_data(){
-    $.ajax({
-        dataType: "json",
-        url: "../data/season_stats.json",
-        success: function (data){
-            preprocessData(data);
-        }
-    });
-};
-
-function preprocessData(data){
-    let dates = [],
-        places = [],
-        topScorerObj = data[0]["best_scorer"],
-        playerAttendanceObj = data[0]["attendance"];
-
-    let goalsLabels = ["GF", "GA", "GD"],
-        goalsData = [data[0]["GF"], data[0]["GA"], data[0]["GD"]];
-
-    data.reverse();
-    for(let item of data){
-        dates.push(item["date_start"]);
-        places.push(item["place"]);
-    }
-
-    let topScorerLabelsAndData = getLabelsAndData(topScorerObj),
-        topScorerLabels = topScorerLabelsAndData[0],
-        topScorerData = topScorerLabelsAndData[1];
-    let maxGoals = Math.max(...topScorerData);
-
-    let attendanceLabelsAndData = getLabelsAndData(playerAttendanceObj),
-        attendanceLabels = attendanceLabelsAndData[0],
-        attendanceData = attendanceLabelsAndData[1];
-
-    let maxAttendance = Math.max(...attendanceData);
-
-    generateTeamPerformanceChart(places, dates);
-    generateTopScorerChart(topScorerLabels, topScorerData, maxGoals);
-    generateAttendanceChart(attendanceLabels, attendanceData, maxAttendance);
-    generateTeamGoalsChart(goalsLabels, goalsData);
-};
-
-function getLabelsAndData(obj){
-    let objArr = [],
-        labelsArray = [],
-        dataArray = [];
-
-    Object.entries(obj).forEach(
-        ([key, value]) => objArr.push([key, value])
-    );
-
-    let objArrSorted = objArr.sort((a, b) => {
-        return b[1] - a[1];
-    });
-
-    for(let item of objArrSorted){
-        labelsArray.push(item[0]);
-        dataArray.push(item[1]);
-    }
-
-    return [labelsArray, dataArray];
-};
-
-function generateTeamPerformanceChart(places, dates){
+export function generateTeamPerformanceChart(places, dates, points){
     let myChart = $("#team_perf");
     let teamPerformanceChart = new Chart(myChart, {
         type: "line",
@@ -73,17 +10,32 @@ function generateTeamPerformanceChart(places, dates){
                 fill: false,
                 borderColor: "#3399FF",
                 lineTension: 0,
-                // pointRadius: 5,
                 pointRadius: function(context){
                     let index = context.dataIndex;
                     return index === 8 ? 9 : 5;
                 },
+                pointHoverRadius: 10,
                 borderDash: [5, 15],
-                // pointBackgroundColor: "#ff3410"
                 pointBackgroundColor: function(context){
                     let index = context.dataIndex;
                     return index === 8 ? "#8F2CFF" : "#ff3410";
-                }
+                },
+                yAxisID: 'y'
+            },
+            {
+                label: "Points",
+                data: points,
+                fill: false,
+                lineTension: 0,
+                borderWidth: 1,
+                borderColor: "#ff3410",
+                backgroundColor: "#64ff25",
+                pointRadius: function(context){
+                    let index = context.dataIndex;
+                    return index === 8 ? 9 : 5;
+                },
+                pointHoverRadius: 10,
+                yAxisID: 'y1'
             }]
         },
         options: {
@@ -116,12 +68,14 @@ function generateTeamPerformanceChart(places, dates){
                 }],
 
                 yAxes: [{
+                    id: 'y',
+                    type: 'linear',
                     display: true,
+                    position: 'left',
                     ticks: {
                         reverse: true,
                         suggestedMin: 1,
                         suggestedMax: 10
-
                     },
 
                     scaleLabel: {
@@ -129,7 +83,23 @@ function generateTeamPerformanceChart(places, dates){
                         labelString: "Ranking",
                         fontSize: 14
                     }
-                }]
+                },
+                {
+                    id: 'y1',
+                    type: 'linear',
+                    display: true,
+                    position: 'right',
+                    gridLines: {
+                        drawOnChartArea: false
+                    },
+                    scaleLabel: {
+                        display: true,
+                        labelString: "Points",
+                        fontSize: 14
+                    }
+                }],
+
+
             },
 
             tooltips: {
@@ -153,7 +123,7 @@ function generateTeamPerformanceChart(places, dates){
     });
 };
 
-function generateTopScorerChart(topScorerLabels, topScorerData, maxGoals){
+export function generateTopScorerChart(topScorerLabels, topScorerData, maxGoals){
     let myChart = $("#top_scorer");
     let topScorerChart = new Chart(myChart, {
         type: "horizontalBar",
@@ -163,13 +133,18 @@ function generateTopScorerChart(topScorerLabels, topScorerData, maxGoals){
                 label: "Goals",
                 display: false,
                 data: topScorerData,
-                // fill: false,
-                backgroundColor: ["#3399FF", "#8000ff", "#ff0040", "#00ffff", "#ff8000", "#ffff00",
-                                  "#c679ff", "#ff2a97", "#ffa51c", "#fffb21", "#54FF18", "#5DFFA4"],
+                fill: false,
+                // backgroundColor: ["#3399FF", "#8000ff", "#ff0040", "#00ffff", "#ff8000", "#ffff00",
+                //                   "#c679ff", "#ff2a97", "#ffa51c", "#fffb21", "#54FF18", "#5DFFA4"],
+                backgroundColor: ["rgb(51, 153, 255, 0.5)", "rgb(128, 0, 255, 0.5)", "rgb(255, 0, 64, 0.5)",
+                                  "rgb(0, 255, 255, 0.5)", "rgb(255, 128, 0, 0.5)", "rgb(255, 255, 0, 0.5)",
+                                  "rgb(198, 121, 255, 0.5)", "rgb(255, 42, 151, 0.5)", "rgb(255, 165, 28, 0.5)",
+                                  "rgb(255, 251, 33, 0.5)", "rgb(84, 255, 24, 0.5)", "rgb(93, 255, 164, 0.5)"],
                 borderWidth: 0.5,
                 borderColor: "#777",
                 hoverBorderWidth: 1,
                 hoverBorderColor: "#000",
+                alpha: 0.5
             }]
         },
         options: {
@@ -211,7 +186,7 @@ function generateTopScorerChart(topScorerLabels, topScorerData, maxGoals){
     });
 };
 
-function generateAttendanceChart(attendanceLabels, attendanceData, maxAttendance){
+export function generateAttendanceChart(attendanceLabels, attendanceData, maxAttendance){
     let myChart = $("#player_attendance");
     let playerAttendanceChart = new Chart(myChart, {
         type: "bar",
@@ -221,8 +196,10 @@ function generateAttendanceChart(attendanceLabels, attendanceData, maxAttendance
                 label: "Attended",
                 display: false,
                 data: attendanceData,
-                backgroundColor: ["#3399FF", "#8000ff", "#ff0040", "#00ffff", "#ff8000", "#ffff00",
-                                  "#c679ff", "#ff2a97", "#ffa51c", "#fffb21", "#54FF18", "#5DFFA4"],
+                backgroundColor: ["rgb(51, 153, 255, 0.5)", "rgb(128, 0, 255, 0.5)", "rgb(255, 0, 64, 0.5)",
+                                  "rgb(0, 255, 255, 0.5)", "rgb(255, 128, 0, 0.5)", "rgb(255, 255, 0, 0.5)",
+                                  "rgb(198, 121, 255, 0.5)", "rgb(255, 42, 151, 0.5)", "rgb(255, 165, 28, 0.5)",
+                                  "rgb(255, 251, 33, 0.5)", "rgb(84, 255, 24, 0.5)", "rgb(93, 255, 164, 0.5)"],
                 borderWidth: 0.5,
                 borderColor: "#777",
                 hoverBorderWidth: 1,
@@ -276,7 +253,7 @@ function generateAttendanceChart(attendanceLabels, attendanceData, maxAttendance
     });
 };
 
-function generateTeamGoalsChart(goalsLabels, goalsData){
+export function generateTeamGoalsChart(goalsLabels, goalsData){
     let myChart = $("#team_goals");
     let teamGoalsChart = new Chart(myChart, {
         type: "doughnut",
@@ -307,5 +284,3 @@ function generateTeamGoalsChart(goalsLabels, goalsData){
         }
     });
 };
-
-load_data();
